Fixed bug where incorrect image rendered when excessive callbacks to bindView were made.
Found that bindView was getting called excessively for ListView
and suspected this was causing rendering problems when the ImageView
gets recycled from ViewHolder and a new bitmap needs to be displayed.

Learned that it is important to use:
- RelativeLayout (not LinearLayout) to contain ListView (I was)
- use 'match_parent' (not wrap_content) for layout height and width

Chaning the the layout height and width to match_parent was the
miracle fix to both the excessive callbacks to bindView and the
incorrect image rendering on the first list item.

http://stackoverflow.com/questions/3859556/cursoradapter-calling-bindview-on-every-item
https://androiddiaries.wordpress.com/2012/06/12/relativelayout-listview-lazy-loading-chaos/

Wonder if the new RecylcerView fixes any of these issues?